# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventGenre = typeof EventGenre[keyof typeof EventGenre];


export const EventGenre = {
  task: 'task',
  shopping: 'shopping',
} as const;

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  detail?: string | null;
  genre: EventGenre;
  done: boolean;
  /**
     * Left border accent color (e.g. "orange" for urgent, "pink" for school)
     * @nullable
     */
  accent_color?: string | null;
  created_at?: string;
}

export type EventInputGenre = typeof EventInputGenre[keyof typeof EventInputGenre];


export const EventInputGenre = {
  task: 'task',
  shopping: 'shopping',
} as const;

export interface EventInput {
  /** @minLength 1 */
  title: string;
  detail?: string;
  genre?: EventInputGenre;
  accent_color?: string;
}

export type EventUpdateGenre = typeof EventUpdateGenre[keyof typeof EventUpdateGenre];


export const EventUpdateGenre = {
  task: 'task',
  shopping: 'shopping',
} as const;

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  detail?: string;
  genre?: EventUpdateGenre;
  done?: boolean;
  accent_color?: string;
}

export interface DailyConfig {
  id: number;
  /**
     * School dismissal time (e.g. "17:00")
     * @nullable
     */
  dismissal_time: string | null;
  has_lunchbox: boolean;
  has_star: boolean;
  has_dog: boolean;
  has_spiral: boolean;
}

export interface DailyConfigUpdate {
  dismissal_time?: string;
  has_lunchbox?: boolean;
  has_star?: boolean;
  has_dog?: boolean;
  has_spiral?: boolean;
}

export interface FridgeSummary {
  active_tasks: number;
  done_tasks: number;
  shopping_items: number;
}