# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digital Fridge API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Fridge events (tasks and shopping items)
  - name: config
    description: Daily configuration (dismissal time, lunch status, event icons)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      description: Returns all tasks and shopping items
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      description: Creates a task or shopping item. If the title contains the Japanese keyword "買う", genre is set to "shopping" automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error

  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      description: Update fields of an event (e.g. toggle done status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Event not found

  /config:
    get:
      operationId: getConfig
      tags: [config]
      summary: Get today's daily configuration
      description: Returns dismissal time, lunch status, and icon flags
      responses:
        "200":
          description: Daily config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyConfig"
    patch:
      operationId: updateConfig
      tags: [config]
      summary: Update today's daily configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyConfigUpdate"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyConfig"

  /summary:
    get:
      operationId: getSummary
      tags: [events]
      summary: Get fridge summary
      description: Returns counts of active tasks, done tasks, and shopping items
      responses:
        "200":
          description: Summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FridgeSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Event:
      type: object
      required: [id, title, genre, done]
      properties:
        id:
          type: integer
        title:
          type: string
        detail:
          type: ["string", "null"]
        genre:
          type: string
          enum: [task, shopping]
        done:
          type: boolean
        accent_color:
          type: ["string", "null"]
          description: Left border accent color (e.g. "orange" for urgent, "pink" for school)
        created_at:
          type: string
          format: date-time

    EventInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        detail:
          type: string
        genre:
          type: string
          enum: [task, shopping]
        accent_color:
          type: string

    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        detail:
          type: string
        genre:
          type: string
          enum: [task, shopping]
        done:
          type: boolean
        accent_color:
          type: string

    DailyConfig:
      type: object
      required: [id, dismissal_time, has_lunchbox, has_star, has_dog, has_spiral]
      properties:
        id:
          type: integer
        dismissal_time:
          type: ["string", "null"]
          description: School dismissal time (e.g. "17:00")
        has_lunchbox:
          type: boolean
        has_star:
          type: boolean
        has_dog:
          type: boolean
        has_spiral:
          type: boolean

    DailyConfigUpdate:
      type: object
      properties:
        dismissal_time:
          type: string
        has_lunchbox:
          type: boolean
        has_star:
          type: boolean
        has_dog:
          type: boolean
        has_spiral:
          type: boolean

    FridgeSummary:
      type: object
      required: [active_tasks, done_tasks, shopping_items]
      properties:
        active_tasks:
          type: integer
        done_tasks:
          type: integer
        shopping_items:
          type: integer