# lib/api-zod/src/generated/types/dailyConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyConfig {
  id: number;
  /**
     * School dismissal time (e.g. "17:00")
     * @nullable
     */
  dismissal_time: string | null;
  has_lunchbox: boolean;
  has_star: boolean;
  has_dog: boolean;
  has_spiral: boolean;
}