# lib/api-zod/src/generated/types/dailyConfigUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyConfigUpdate {
  dismissal_time?: string;
  has_lunchbox?: boolean;
  has_star?: boolean;
  has_dog?: boolean;
  has_spiral?: boolean;
}