# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventGenre } from './eventGenre';

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  detail?: string | null;
  genre: EventGenre;
  done: boolean;
  /**
     * Left border accent color (e.g. "orange" for urgent, "pink" for school)
     * @nullable
     */
  accent_color?: string | null;
  created_at?: Date;
}