# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventInputGenre } from './eventInputGenre';

export interface EventInput {
  /** @minLength 1 */
  title: string;
  detail?: string;
  genre?: EventInputGenre;
  accent_color?: string;
}