# lib/api-zod/src/generated/types/eventUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventUpdateGenre } from './eventUpdateGenre';

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  detail?: string;
  genre?: EventUpdateGenre;
  done?: boolean;
  accent_color?: string;
}