# lib/api-zod/src/generated/types/fridgeSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Fridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FridgeSummary {
  active_tasks: number;
  done_tasks: number;
  shopping_items: number;
}